refactor: thin __init__.py files
