+++ README.md
@@ -36,8 +36,6 @@ This can be found in https://discord.com/applications/{application_id}/bot
 
 `TOKEN=DISCORD_BOT_TOKEN`
 
-Intent?
-
 ### Docker Deployment
 
 Build image
